Fixing the whole structure of the app.
