+++ lib/data/repositories/firebase_auth_repository.dart
@@ -1,0 +1,79 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:google_sign_in/google_sign_in.dart';
+import 'auth_repository.dart';
+
+class FirebaseAuthRepository implements AuthRepository {
+  final FirebaseAuth _firebaseAuth;
+  final GoogleSignIn _googleSignIn;
+
+  FirebaseAuthRepository({
+    FirebaseAuth? firebaseAuth,
+    GoogleSignIn? googleSignIn,
+  })  : _firebaseAuth = firebaseAuth ?? FirebaseAuth.instance,
+        _googleSignIn = googleSignIn ?? GoogleSignIn();
+
+  @override
+  Future<User?> signInWithEmailAndPassword({
+    required String email,
+    required String password,
+  }) async {
+    try {
+      final UserCredential credential =
+          await _firebaseAuth.signInWithEmailAndPassword(
+        email: email,
+        password: password,
+      );
+      return credential.user;
+    } on FirebaseAuthException catch (e) {
+      throw Exception('Sign in failed: ${e.message}');
+    } catch (e) {
+      throw Exception('An unknown error occurred during sign in.');
+    }
+  }
+
+  @override
+  Future<User?> signUpWithEmailAndPassword({
+    required String email,
+    required String password,
+  }) async {
+    try {
+      final UserCredential credential =
+          await _firebaseAuth.createUserWithEmailAndPassword(
+        email: email,
+        password: password,
+      );
+      return credential.user;
+    } on FirebaseAuthException catch (e) {
+      throw Exception('Sign up failed: ${e.message}');
+    } catch (e) {
+      throw Exception('An unknown error occurred during sign up.');
+    }
+  }
+
+  @override
+  Future<User?> continueWithGoogle() async {
+    try {
+      // Create a new Google provider
+      final GoogleAuthProvider googleProvider = GoogleAuthProvider();
+
+      // Sign in with a popup and wait for the result
+      final UserCredential userCredential =
+          await FirebaseAuth.instance.signInWithPopup(googleProvider);
+
+      // Return the user
+      return userCredential.user;
+    } on FirebaseAuthException catch (e) {
+      throw Exception('Google sign in failed: ${e.message}');
+    } catch (e) {
+      throw Exception('An unknown error occurred during Google sign in.');
+    }
+  }
+
+  @override
+  Future<void> signOut() async {
+    await Future.wait([
+      _firebaseAuth.signOut(),
+      _googleSignIn.signOut(),
+    ]);
+  }
+}
